requirements.txt
blinker==1.9.0
click==8.3.3
feedparser==6.0.12
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Markdown==3.10.2
MarkupSafe==3.0.3
PyMySQL==1.1.3
python-dotenv==1.2.2
python-frontmatter==1.1.0
PyYAML==6.0.3
sgmllib3k==1.0.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
Werkzeug==3.1.8
